feat: make request to Product - get, post, patch, delete
