+++ sprint_4/index.html
@@ -7,6 +7,6 @@
 </head>
 <body>
     <h1>Sprint 4</h1>
-    <script src="main.js"></script>
+    <script type="module" src="main.js"></script>
 </body>
 </html>
