+++ sprint_4/main.js
@@ -1,90 +1,34 @@
-async function getArticleList(page=1, pageSize=10, keyword="") {
-  const url = `https://sprint-mission-api.vercel.app/articles?page=${page}&pageSize=${pageSize}&keyword=${keyword}`;
-  try {
-    const response = await fetch(url);
-    if (!response.ok) {
-      throw new Error(`Response status: ${response.status}`);
-    }
+import { getArticleList, getArticle, createArticle, patchArticle, deleteArticle} from './ArticleService.js';
+import { getProductList, getProduct, createProduct, patchProduct, deleteProduct } from './ProductService.js';
 
-    const json = await response.json();
-    console.log(json);
-  } catch (error) {
-    console.error(error.message);
-  }
-}
+// const data = {
+//   id: 586,
+//   name: "zino patch test",
+//   description: "edit test",
+//   price: 99900,
+//   manufacturer: "edit zino",
+//   tags: ["what tag"],
+//   images: ["image url"]
+// }
 
-async function getArticle(id) {
-  const url = `https://sprint-mission-api.vercel.app/articles/${id}`;
-  try {
-    const response = await fetch(url);
-    if (!response.ok) {
-      throw new Error(`Response status: ${response.status}`);
-    }
+// console.log(deleteProduct(586));
+// console.log(getProduct(586));
 
-    const json = await response.json();
-    console.log(json);
-  } catch (error) {
-    console.error(error.message);
-  }
-}
+// const data = {
+//   name: "zino test",
+//   description: "test",
+//   price: 1000,
+//   manufacturer: "zino",
+//   tags: ["what tag"],
+//   images: ["image url"]
+// }
 
-async function createArticle(title, content, image) { 
-  const url = "https://sprint-mission-api.vercel.app/articles";
-
-  try {
-    const response = await fetch(url, {
-      method: "POST",
-      headers: {"Content-Type": "application/json",},
-      body: JSON.stringify({title, content, image}),
-    });
-
-    if (!response.ok) {
-      throw new Error(`Failed to create article: ${response.status}`);
-    }
-
-    const result = await response.json(); 
-    console.log("Article created successfully:", result);
-  } catch (error) {
-    console.error("Error:", error.message);
-  }
-}
-
-async function patchArticle(id,title, content, image) { 
-  const url = `https://sprint-mission-api.vercel.app/articles/${id}`;
-
-  try {
-    const response = await fetch(url, {
-      method: "PATCH",
-      headers: {"Content-Type": "application/json",},
-      body: JSON.stringify({title, content, image}),
-    });
-
-    if (!response.ok) {
-      throw new Error(`Failed to patch article: ${response.status}`);
-    }
-
-    const result = await response.json(); 
-    console.log("Article patched successfully:", result);
-  } catch (error) {
-    console.error("Error:", error.message);
-  }
-}
-
-async function deleteArticle(id) { 
-  const url = `https://sprint-mission-api.vercel.app/articles/${id}`;
-
-  try {
-    const response = await fetch(url, {
-      method: "DELETE",
-    });
-
-    if (!response.ok) {
-      throw new Error(`Failed to delete article: ${response.status}`);
-    }
-
-    const result = await response.status; 
-    console.log("Article deleted successfully:", result);
-  } catch (error) {
-    console.error("Error:", error.message);
-  }
-}
+// const data = {
+//   id: 586,
+//   name: "zino test",
+//   description: "test",
+//   price: 1000,
+//   manufacturer: "zino",
+//   tags: ["what tag"],
+//   images: ["image url"]
+// }
